arcappcompat: Update tast test for Disney
It is observed that on some Chromebooks, the app
launches in tablet mode and the log out button
is not visible. Need to scroll the page to make
the log out button visible. Added a code to
Handle it. Also observed that clicking on the
logout button for a single time does not click
the logout button. Added a common func in
loginutil to handle the scenario.

BUG=b:242607641
TEST=tast run <kohaku> arcappcompat.Disney.vm_clamshell_mode
tast run <kohaku> arcappcompat.Disney.vm_tablet_mode
tast run <kevin> arcappcompat.Disney.clamshell_mode
tast run <kevin> arcappcompat.Disney.tablet_mode

Change-Id: I061c8897784eb3ea73d015b40faa1785c3015e9c