tast-tests: Switch the SWA on by default in tast
The Files SWA is on by default as of M106, this flips the default
behaviour in tast to launch the SWA by default and the Chrome app as an
opt in behaviour.

NOTE: The policy.PinnedLauncherApps relies on the Files chrome app to be
pinned by default. When changing to SWA the existing signal (absence of
the "Unpin" button) is no longer correct. To ensure the framework
enables SWA by default this inverts the logic to have
policy.PinnedLauncherApps run the chrome app whilst we investigate the
reasons SWA doesn't behave the same.

BUG=b:242940102, b:242930307, b:243619838
TEST=tast run DUT filemanager.* crostini.*

Change-Id: Ief267536dfceaef1a5660d443ce9512f7c434f5c