tast-tests: Add basic arc.LinkCapturing test
This basic test scaffold verifies that clicking a link in the browser
which can be handled by an ARC app does not automatically open that app.
This is an initial simple test scenario which will be expanded in
follow-up CLs to also test links clicked in ARC, and with the "Opening
supported links" app setting enabled/disabled.

BUG=chromium:1353271
TEST=tast run $DUT arc.LinkCapturing

Change-Id: I2c2abc5f283d7c3f20c614deca00e23aa493e110