tast-tests: Improve stability of enterprise.ARCProvisioning
In this change we're creating a separate retryIfStable method which is
only used for the temporary retries added for test stability while
the normal retry method will remain when the unstable version of the
test is removed.

Removed the Chrome crash detection code which no longer occurs on the
unstable version.

Increased the timeout for unstable version since apps fail to install
within the 10-minute time limit.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning{.vm|.vm_unstable|unstable}

Change-Id: Id0647cd1dfddbd2a882e63bb92926157f1a7d207