tast-tests: another attempt to fix apps.LaunchImageInPhotosFromGallery
The test is still flaky after https://crrev.com/c/3805245. According to the fail log, ACUITI failed to take stable screen after click button, probably due to animation or other screen changes.

When prompts shown up, it is intended to click button to close it. LeftClick already ensures the screen is stable, so it is okay to take immediate screenshot to wait for the prompt gone.

Tested it on a leased hana device, 20/20 pass.

BUG=b:227115550
TEST=tast run <DUT> apps.LaunchImageInPhotosFromGallery*

Change-Id: I7e427e85da11db2c40931277d92785886e6cdafa