Projector: Verify SODA is installed in Tast test
This CL modifies the projector.NewScreencastButtonCondition test to
assert that SODA is installed instead of checking for the appropriate
error message and exiting prematurely.

Note that this test no longer passes in VM because SODA is not
available in VM, but that's okay because we don't run this test against
VM by default.

BUG=b:242316782
TEST=tast run -buildbundle=cros localhost:2222
projector.NewScreencastButtonCondition

Change-Id: I00440b1c71d27034d841339b4a5a4776f7619d0a