tast-test: Trigger an update for recent apps on homecher
Fix some of the failures in tablet mode by triggering an update in the
homcher after setting up the test so that RecentApps is able to update
and show on some of the devices.

For the CWS app, sometimes it will open a window twice, which will
result in an error as there are multiple matching windows for the
WaitUntilExistCondition. For this case, just select the first one to
avoid the Fatal.

Tested on a Dru device that was consistently failing.

BUG=chromium:1344547
TEST=tast run <DU>T launcher.RecentApps.*

Change-Id: I7fe683eb07e37cdcbbb74913a4790a762078ff41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3838868
Reviewed-by: David Welling <davidwelling@google.com>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Tested-by: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>