tast-tests: [Launch Lacros at login screen] Consider /var/log/lacros
This commit is part of a series, meant to provide the capability
of pre-launching Lacros at login screen for performance reasons.
Design doc: http://go/lacros-loginscreen-design

Previous commits:
- http://crrev/c/3747600 (Introduce proxy class BrowserParamsProxy)
- http://crrev/c/3750558 (Introduce BrowserPostLoginParams)
- http://crrev/c/3787977 (Create pre-login Lacros log directory)

When launching Lacros at login screen, we need a directory outside
of the cryptohome to store logs before login.
This change fixes the security.StatefulFile test to check the
permissions of this directory (/var/log/lacros).

BUG=1363933
TEST=tast tests

Change-Id: I20231a2991fd03df5a157f18ea8347e2294464e3