Revert "[CrOS SLP] Add tast tests for when Lock SIM setting is off."
This reverts commit 22f19c8f4bce685e7189206ebde8673f4e85bd1e.

Reason for revert: See b/243121258. This CL updates defs.go to
introduce AllowCellularSimLock, which is causing build issues when
going through the uprev process because the field is removed again
by the sync script. This file is in sync with policy_templates.json
and should not be updated by hand or with local changes to
policy_templates.json.

We need to revert https://crrev.com/c/3806332 as well because it
also uses this field.

Original change's description:
> [CrOS SLP] Add tast tests for when Lock SIM setting is off.
>
> BUG=b:228093904
> TEST=network.SimLockPolicyLockSettingOff
>
> Change-Id: Id58bc8a7bd334e8eb2852be485fb26b3990ffbd0
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3792748
> Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
> Commit-Queue: Regan Hsu <hsuregan@chromium.org>
> Tested-by: Regan Hsu <hsuregan@chromium.org>

BUG=b:228093904, b:243121258
TEST=../tast/fast_build.sh -T

Change-Id: Idf91d5448b0fd1b197595c4234202c2b3b1c5ec6