accountmanager: Fix "failed to click More actions button"
Inside the for-loop in TestCleanup we call
`ui.WaitUntilGone(accountMoreActionsButton)` and right after that check
for the next button with `ui.IsNodeFound(ctx, moreActionsButton)`.
Unfortunately, `ui.IsNodeFound` can return true for the same element
sometimes even after `ui.WaitUntilGone` completed.

Fix this by reloading the window (navigating to
chrome://os-settings/accoutnManager) inside the for-loop to make sure
the UI is updated.

BUG=b:237655468, b:243124200
TEST=tast run <ip> accountmanager*

Change-Id: I1ab4d76b8912f70f7d7aa5d0955f8b912f9bc9c9