tast-tests: Switch ListAuthFactors to using binary-protobuf output
This changes the implementation of the "call ListAuthFactors and parse
the output" to use the new binary-protobuf output format, and
proto.Unmarshal instead of manual string parsing.

This also fixes a bug in one of the error messages, where the "got" and
"want" were reversed for the number of auth factors reported.

Note that this will be helpful for parsing the supported auth factors
output, which is harder to parse by hand.

BUG=b:242228418,b:227729894
TEST=tast test <dut> cryptohome.AddRemoveFactors.*

Change-Id: Ic342ac98a480b2af56859eba3ef4248edcc8640f