platform: add test for non-empty form-factor
We recently found that the form-factor was not set in the model.yaml
for nautilus. This caused some tests to be skipped as the device did not
satisfy the `hwdep.FormFactor(hwdep.Convertible)` requirement.

Add a test to find any other devices with missing form-factors.

We may want to integrate this test into the fixture hwdep logic in the
future. For now though, it is a seperate test to avoid breaking the test
fixture.

BUG=b:242094716
TEST=platform.FormFactor

Change-Id: I6827b31a8a82a0f33b0da1b65a14f9e73086b8fe