network: Reset StaticIPConfig in RoutingIPv4Static* tests
This patch adds a defer function to reset the StaticIPConfig property in
the test, to avoid a racing issue inside shill (see
b/239753191#comment8). This issue does not affect these two tests, but
may put the physical interfaces on DUT in a bad state which could
potentially affect the following Routing tests.

BUG=b:239753191
TEST=../platform/tast/fast_build.sh -T
TEST=Ran `tast run $DUT network.RoutingIPv4Static*`, logs indicated that
  shill no longer tried to install static IP used in the test to eth0.

Change-Id: I22beb2263d007b7f58fe1f44468908aa46740f58