tast-tests: Add supported factors coverage to the AddRemoveFactors test
This doesn't modify the overall test flow, it just adds more to the
expectations. In addition to checking the configured factors, the test
now also checks the supported factors.

The expectations aren't quite as simple because we need to look at some
of the test and hardware parameters to decided exactly what the expected
factors are, but they're still pretty basic to compute.

BUG=b:227729894
TEST=tast test <dut> cryptohome.AddRemoveFactors.*
TEST=tast unit tests passed

Cq-Depend: chromium:3833556
Change-Id: I673dfc8c1e5bbc513bd37504304ac4875b22f92a