tast-tests: Add helper for fetching available cellular network information.
This adds a helper "cellular.NetworkProvider" which provides
methods for fetching general cellular network metadata for both eSIM and pSIM
cellular networks. Currently, this is focused around network names and
includes methods to bulk query/update cellular eSIM network names but can
later be extended to include additional network properties.

This implementation is distinct from the mojo implementation of NetworkState
in the netconfig package to provide access to the underlying D-BUS
objects to be used with cellular.Helper and other support items.

BUG=b:242318523
TEST=tast run <DUT> network.CelularPolicyConnection(changes not included)

Change-Id: I293a52c684cab81daaf0b80c8da4cbbd8bf3c6ce