tast: Update firmware_ECWakeSource test.
Move stable tests to stable suite.
The keypress test expects different keyboard behaviour on devices that
support the `tabletmode` command but it appears toggling `tabletmode` is
not reflecting in the behaviour of the internal keyboard so this
additional case was removed as it was causing a significant amount of
failures.

BUG=b:194908279
TEST=firmware.ECWakeSource.*

Change-Id: If15a4f28c20b9663855b260e5abd8f146d0c21b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3849962
Tested-by: Kshitij Shah <tij@google.com>
Commit-Queue: Kshitij Shah <tij@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>