TestALSACommand: Change error reporting method.
It is recommended not to pass testing.State to function as an argument.
Currently, audio.DevicePlay and audio.DeviceRecord calls function
device.TestALSACommand. Both of them pass in testing.State. Among them,
the passing of testing.State in device.TestALSACommand is clearly not
needed, since it only uses it to report errors.

Modify the arguments and return value for device.TestALSACommand to
report errors by returning an error value instead of passing
testing.State.

BUG=b:241813621
TEST=tast run $DUT audio.DevicePlay. Check error message when there are
no sound cards.
TEST=tast run $DUT audio.DeviceRecord. Check error message when there
are no sound cards.

Change-Id: I85da896f7fbd1eec809eda41e64d39a43d58ce83