mta: Fix screenshot's size mismatching issue in wmp.FullScreenshot
The test wmp.FullScreenshot all fails on the board strongbad and kukui.
Because the orientation of the screenshot is different from the natural
orientation of the DUT.
This CL fixes the issue by swapping the WidthInNativePixels and
HeightInNativePixels value to match the current screen ui orientation
when orientation angle is 90 or 270 degrees.
This CL also removes the unnecessary actions because the actions doesn't
help to change the screenshot's size or DUT's native setting.

BUG=b:243342186
TEST=tast run <DUT> wmp.FullScreenshot

Change-Id: Ibb10feaf16288bf796e8d711e4acad4cb2c86c09