windowarrangementcuj: Detect when an app crashes
If a window vanishes during a drag in ui.WindowArrangementCUJ.*, the
test will fail with an error giving the title of the vanishing window.
This may give insight into high failure rates of the tablet variants, as
I suspect most of the failures are caused by the ARC app crashing.

BUG=b:243533353
TEST=tast run <DUT> ui.WindowArrangementCUJ.*

Change-Id: I8378af18419e53d1229fc285e6454906cc102760