tast-tests: Workaround to set tablet mode for lacros
There seems to be an issue that Lacros does not start in tablet mode if
it has not been opened before the tablet mode is turned on.
It is unsure if it would happen only in Tast environment not in
production, but it would be safe to get around this issue without test
coverage loss by opening browser windows first then setting the tablet
mode later.

BUG=chromium:1348824, b:243698545
TEST=tast run $DUT ui.DragWindowFromShelfPerf*

Change-Id: Id2924559d1a6b45e229a6013d66892c5a97f6b54