faft: Improve on ECVerifyVK
Extend timeout in waiting for the virtual keyboard,
and restore DUT's tablet mode setting before the test
ends if it was changed. Also, saving logs on error seems
to be preventing screen record from being saved, and
successful exit of the test. Temporarily disable
saveLogsOnError until the next bug fix.

BUG=b:200305355
TEST=manual on Eldrid, Storo360, Coachz, and Dumo

Change-Id: Ie47e77df741c6b8bd9d25e0316297bd814607027