fingerprint: Disable DBus bus actiavtion for fwupd service
DBus bus activation is a mechanism which is responsible for starting
service on connection attempt. It turns out that the mechanism is used
to start fwupd service when Chrome is trying to call some methods from
org.freedesktop.fwupd service.

To work properly, fwupd enables powerd service. As a result powerd can
be running despite the fact that it was disabled at the beginning of the
test.

The problem is fixed by moving 'org.freedesktop.fwupd.service' file to
/mnt/stateful_partition at the beginning of the test. It will prevent
fwupd and powerd from starting during the test.

BUG=b:243645924
TEST=Run 'tast run <IP> firmware.FpReadFlash' on brya.
TEST=TAST unit test passed.

Change-Id: Ia4a6d3da05331344ed854255952e6b06b5f14143