cellular: Fix multisim test for MBIM modems
Shill does not export services for the inactive eSIM slot if the iccid
of the eSIM slot is unknown. Chrome uses Hermes to switch slots in this
case. For inactive pSIMs, Shill exports a service  with
iccid="unknown-iccid", so that Chrome can connect to the pSIM.
See go/dual-sim-mbim for details.

BUG=b:200813616
TEST=tast run $cr cellular.ShillCellularSimSlots

Change-Id: I6127c6930aefc4080ed74cd49994d8138b3d333c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3857604
Commit-Queue: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>