tast-tests: Add btpeers device type support and discovery test
This change adds a new EmulatedBTPeerDevice struct that tests
can use to easily configure a btpeer to act as any supported
bluetooth device and cache related key data about it. It also
adds a new test, DiscoverEmulatedBTPeerDevices, which
demonstrates this new support and also validates that a DUT
can discover the emulated device as expected.

Testing of EmulatedBTPeerDevice found a few gaps in both the
Chameleond XMLRPC interface and in the XMLRPC client library,
so both have been updated accordingly.

BUG=b:217633791
TEST=tast unit tests passed; Ran new test DiscoverEmulatedBTPeerDevices
successfully with two btpeers

Change-Id: If0174f09d4e278f14efe34d2d1c086391a9d682d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3857487
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>