printer.Print: Clear notifications before selecting printer
If a device has multiple real printers plugged in, the resulting
notifications may cover the printer name in the destination dialog.
This causes the notification to steal the click instead of selecting the
desired virtual printer.  This probably doesn't happen in the labs, but
it comes up when running on developer desktops.  Clear the notifications
before trying to select the printer to ensure the number of attached
printers doesn't matter.

BUG=None
TEST=tast run printer.Print with 7 attached USB printers

Change-Id: I3a04118fb1249cb7eb83071153abc39733f6933d