tast: Add previous BSSID into ignore list in BSSTM test
This CL updates wifi.BSSTMRequest to add roamed-from BSSID (that is, the
BSSID from which the BSSTMRequest was sent) into the ignore list on the
DUT using the wpa_cli. Importantly, this ignore list is not used to
trigger any roam decisions, so adding a BSSID into the ignore list will
not cause the DUT to move away from that BSSID. Instead, it simply
prevents the DUT from roaming back to that BSSID after it has moved
away.

This means that this change is not invasive to the core behavior
under tested by this test, which is that the DUT must roam away from a
BSS which has sent a BSSTMRequest. There is currently no platform
expactation that a DUT will actually *stay* connected to its new BSS
after transitioning due to a BSSTM.

The BSSID is added to the ignore list twice in order to ensure that it
remains in the ignore list for long enough. Each wpa_cli invocation
actually results in the BSSID being added twice, so the two wpa_cli
invocations in this test actually translate to 4 insertions from
wpa_supplicant's perspective. This results in an ignore list duration of
120 seconds, which should be plenty.

This change should reduce or eliminate a specific category of flake
within this test, which is caused when the device roams back to the
previous BSSID some time after it initially roamed away.

BUG=b:229891363,b:229891062
TEST=run wifi.BSSTMRequest 10 times on Bob with and without this change.
The test passed 8/10 times without this change, and 10/10 times with
this change.

Change-Id: I9200e437e90d22e2930846ecc61a0a0b375973cf