tast-tests: LacrosVariantUnneeded to ui.CUJRunner
This test seems to have been in experiment and disabled for a while.
It is safe to mark as LacrosVariantUnneeded until reenabled.

At a glance, cujrunner.runActionOpenURL is the only part that may
possibly need to be migrated to lacros when needed.

BUG=b:216242056
TEST=tast run $DUT ui.CUJRunner

Change-Id: I1876f7a8c23f497780ff1e05b08315666e1630e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3858617
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Auto-Submit: Hyung Tae Kim <hyungtaekim@chromium.org>