tast-tests: Simplify arc.UnicornBlockedApps
We don't need to serve a policy with blocked apps in it since Play Store
does not rely on ARC policy to know which apps are blocked.

Added a check to wait for provisioning before launching Play Store.

Added the DisableSyncFlags flag in fixture to avoid flakiness due to
Play Store updates triggering during the test.

BUG=b:237023275
TEST=tast run DUT arc.UnicornBlockedApps

Change-Id: I547e172804c9a81381ee016a5cd0e30881446b43