ui: Fix ChromePIPRoundedCornersUnderlay failing on nocturne
Fixes a problem with how the tests ui.ChromePIPRoundedCornersUnderlay*
resize the PIP window. The tests use mouse input or touch input,
depending on whether the device defaults to clamshell mode or tablet
mode. The resizing drag begins exactly at the corner of the window
bounds, which works reliably with mouse input but not with touch input.
Now the drag will begin just a tiny distance inset from the corner.

BUG=b:244235065
TEST=tast run <DUT> ui.ChromePIPRoundedCornersUnderlay* on nocturne with
the detachable keyboard/trackpad detached

Change-Id: Ia9bb7a476798f27a83773caa62cb2044f3023e9c