Fix drivefs_pooled_storage tast test
Stop verifying that the visual signal is being
displayed when copying a new file into drive.
This leads to flaky tests since the time for
syncing to happen might exceed the timeout
for finding the corresponding text element of
the visual signal. Also, it leads to significantly
slower tests when we are already testing that
the banners are being displayed correctly.
Since both the banners and the visual signals
emanate from the same source of truth, it might
suffice to just check the banners in this test.

https://tastboard.corp.google.com/test?contact=chromeos-files-syd%40google.com&testname=filemanager.DrivefsPooledStorage*&dep=&attr=

BUG=chromium:1324367
TEST=tast run dut *.DrivefsPooledStorage*

Change-Id: If439c1d0a2cb4b9ada09d936f291a26fa3673250