tast-tests: Verify AuthenticateAuthFactor reply
Parse the response proto from cryptohome's AuthenticateAuthFactor reply
and check that, when the call succeeds, it actually has the
"authenticated" field set to true.

This is preparation change for testing the AuthIntents that will be
returned in the reply protos by cryptohome soon too.

BUG=b:240596931
TEST=tast run $DUT cryptohome.UserSecretStash

Change-Id: Ie05589fd1e2c907f44d7cb780bb160a8633c0622
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3865499
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>