tast-tests: Use DoDefault instead of LeftClick ofr install button
In small screens, sometimes the add button in the webstore is out of
screen bounds, and although the accessibility tree is able to find it,
it is unable to click it.

To ensure that the install action is triggered every time, use the
DoDefault method in ui automation, which will trigger a left click on
the button even if it is out of bounds.

BUG=chromium:1344547
TEST=../tast/fast_build.sh -T

Change-Id: Ide9d0648370ca0d272611e7355dae4fcbce20807
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3868128
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Ana Salazar <anasalazar@chromium.org>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>