Turn the webcam on or off depending on the cam field
Some devices appear to be running MeetCUJ with the camera turned off.
Instead of assuming the camera will be turned on, set the camera's on
or off state directly based on the cam field in meetTest.

BUG=b:240172166
TEST=tast run <DUT> ui.MeetCUJ.2p on a dedede/metaknight device, which
has regularly failed due to having the camera turned off.

Change-Id: I69fba92652c17ba0f17e2994b7f81d929b6138d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3867388
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Yichen Zhou <yichenz@chromium.org>