tast tests: Add debug info to rollback helper
EnterpriseRollbackPreviousVersion does not restore the original
image when it calls rollback.RestoreOriginalImage. However, the
PostTest part of the fixture afterwards restores it successfully.

Add logs to show which is the active partition before and after
the restoration step.

The method has also been modified to return if it is not
necessary to restore the image.

BUG=b:241391509
TEST=tast run -var=updateutil.tlwAddress="localhost:7151" DUT
autoupdate.EnterpriseRollbackPreviousVersion.rollback_1_version

Change-Id: Iea2f6d72b7ff89ab04d9f49e8aa09a87fd5b62f6