policy: Add integration test.
Add integration test for ZTE enrollment. It took me a long time to get
this test to pass but this is just the first version, there is still a
lot of work to do. Wanted to get something submitted and then I'll build
on it. This by itself will not be running.

Next steps are:
1. Add deprovisioning to this test
2. Add the curl command for pre-provisioning after the test is
   completed.
3. Figure out how to run the reset script on the test device once test
   is done.

BUG=b:230622059
TEST=https://paste.googleplex.com/5824057596444672

Change-Id: I93348b4df46059ec2e0d98d8524d2be54167fb2b