policy: Add integration test.
Add integration test for ZTE enrollment. It took me a long time to get
this test to pass but this is just the first version, there is still a
lot of work to do. Wanted to get something submitted and then I'll build
on it. This by itself will not be running.

BUG=b:230622059
TEST=ran test on local and lab machines

Change-Id: I93348b4df46059ec2e0d98d8524d2be54167fb2b