featured: Add helper program for tast test
At present there is no feature_library tast test to verify the
correctness of a feature's determined state and parameters.

Add helper program for feature_library tast test to check for
correctness of feature state and parameters. Include parameterized tests
that check when a feature is explicitly enabled and explicitly disabled.

BUG=b:237311752
TEST=built and manually verify output on DUT

Change-Id: Id3ddf118e440c87a447ddc9472cee05096db8eaa