tast-tests: WiFi: Remove the check on beacon signal avg from wifi.Nl80211StationInformation
wifi.Nl80211StationInformation requires device support beacon signal avg
and check if the difference between beacon signal avg and signal, signal
avg is less than a threshold.

This test doesn't guarantee that the value returned by the device are
correct, so that it is unnecessary to check if the the signal strength
makes sense. This CL removes the check on signal strength.

The design doc is available at
go/shill-wifi-nl80211-test

BUG=b:235532246
TEST=tast run -var=router=localhost:2210 -var=pcap=localhost:2220
localhost:2200  wifi.Nl80211StationInformation

Change-Id: I1b8e8cfdf6b40a1057c4b15b87875d18ed1c1095