tast-tests: Reconnect to DUT before health check when needed
Sometimes before a recoverUnhealthyDUT call, earlier DUT reboot
timeouts put the fixture in a state without a live ssh connection
to the DUT. Thus, the recovery steps it tries just fail because
there is no live ssh connection to try them. This adds step
to recover the ssh connection prior to other recovery actions,
potentially giving more opportunities for the fixture to recover
and not just fail tests due to a fixture problem.

BUG=b:239950948
TEST=tast unit tests passed; successfully reconnected during a wifi
test run by manually killing the ssh connection right before
recoverUnhealthyDUT is called and the test (wifi.SimpleConnect.80211a)
was still able to pass as a result.

Change-Id: I2b559e4ec807e700e7bc9e97cbbf5f1e1fbae5c0