crostini: optimize Crostini installer
The current implementation waits for the installation window to be gone
regardless of whether the installation succeeds or fails.

It could hang a test if the installation fails. E.g., if the
installation fails at around 05:00, the test will wait until the timeout
(15 mins), which is a waste of time and prevent retires if the test
runner reaches the global time time.

Revise to use the "installing Linux..." message as the indication of
installing in process, and check the error message if the installation
window still exists after the install message is gone.

This doesn't fix the issue b/239872199, but allows more retries.

BUG=b:239872199
TEST=tast run <DUT> crostini.*

Change-Id: I1af4180167a1ad43939b1261c4605f134359d816