kerberos: Fix kerberos.*Website tests
Pressing "Advanced" and "Proceed" buttons are independent actions and
they shouldn't be in a single cycle(Poll). For example if "Advanced"
succeeds and "Proceed" fails the first time, "Advanced" is gonna be
pressed in the next iteration again, which would actually close the
menu.

However, checking if website title contain 401 or not does depend on
getting the non-empty title and should be in the same Poll. Otherwise
we can get the title of the page before it was loaded, and then check
if it contains 401.

+ minor refactoring.

BUG=chromium:1249427
TEST=tast run localhost:2200 kerberos.*Website

Change-Id: I63313384203885fde13f2cecd90b95e28d3e18bc