kerberos: Fix kerberos.*Website tests
Pressing "Advanced" and "Proceed" buttons are independent actions and
they shouldn't be in a single cycle(Poll). For example if "Advanced"
succeeds and "Proceed" fails the first time, "Advanced" is gonna be
pressed in the next iteration again, which would actually close the
menu.

In other places there is no need to use Poll, since we can just check
for the status of the webpage (if it's loaded or not)

+ minor refactoring.

BUG=chromium:1249427
TEST=tast run localhost:2200 kerberos.*Website

Change-Id: I563613305bc2063ca6c66fdcb191f4769c5ee8bc