firmware_test_handler: Return error when connection to DUT is lost
ReimageFPMCU() function reboots DUT at least once. It can happen that
the test doesn't reconnect to DUT after reboot. If this happens in a
Close() method then error returned by ReimageFPMCU() is saved and
execution continues, because we try to restore other things (services)
though eg. flashing failed.

However if ReimageFPMCU() fails because we can't reconnect to the DUT
then we should return from Close() immediately because there is nothing
to do without connection to DUT.

BUG=b:240299731
TEST=Run 'tast run <IP> firmware.FpSoftwareWriteProtect' on gimble with
     brya-manatee. Make sure the nil pointer dereference doesn't occur.

Change-Id: Id0e622ed22124386e430cc740fbdf149354f9031