ARCVPNConfigs: Clean up VPN connection after disconnect
After we switch the underlying implementation of the vpn connection from
ConnectNamespace API by patchpanel to the virtualnet package, we need to
clean up the vpn connection object before starting a new one. This patch
fix this issue by putting the logic for one connection into one
function, so that cleanup will be executed as a defer function at the
end of the function. This refactor also helps reduce the duplicated
code.

BUG=b:244706389
TEST=`tast run $DUT network.ARCVPNConfigs` on ARC-R device, note that
  the test is failing with another reason on P device now

Change-Id: Id0297c75ca0554f35526d1bc4c5051238ced4136
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3874544
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Cassie Wang <cassiewang@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>