tast-tests: Poll for Files app window and make informational
After launching the Files app it takes a little while for the title to
update with the prefix. It starts as Files then after loading the list
view it ends up as "Files - My files". This adds a 5s poll to wait for
the window to stabilize before maximizing. The test is also currently
flaky, so demote it to informational whilst investigating why smbd is
crashing.

BUG=chromium:1360263
TEST=tast run filemanager.SMB (this still fails due to smbd crashing)

Change-Id: Ieecd78b5036f0d0c631238ba4759492873ff3a9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3874508
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>