login.OfflineLoginWithUsernameAndPhotosDisabled: Use separate cleanupCtx
loginOffline() should create and use its own cleanupCtx instead of use
the captured one from outside, since network.ExecFuncOnChromeOffline()
will also shorten the ctx by itself. In the current code, when
DumpUITreeWithScreenshotOnError() on L81 timeouts with the cleanupCtx,
the network.ExecFuncOnChromeOffline() function has no chance to run its
own cleanup because cleanupCtx lives longer than the ctx passed to
network.ExecFuncOnChromeOffline(), and as a result, the iptables will
not be cleaned up, which blocks the following other tests which will use
the network.

BUG=b:245043762
TEST=../platform/tast/fast_build.sh -T

Change-Id: I0ebde882762c1d04f4986d5101baaa2299237884