webrtc: Add h264 VBR tests to MediaRecorder
MediaRecorder uses VBR encoding if the hardware encoder supports it.
The flag of enabling hardware VBR encoding is not enabled by default.
This CL adds the test of running MediaRecorder enabling the feature
so that VBR encoding is enabled in MediaRecoder test.
The tests will be removed once the feature is enabled by default.

BUG=b:181797390
TEST=webrtc.MediaRecorder*.*vbr

Change-Id: I4d2509942cf61db1a96270ee1ae474734b96fff7