tast-test: Use Chrome app instead of Files for Recent Apps
With a recently introduced change that makes the Files SWA app launched
by default, it is safer to stop relyig on the files app tp interchange
the behaviour and it makes more sense to use the Chrome app.

This CL also closes the newly installed app after opening the first
since it was causing some crashes to drag the shelf, particularly for
the Google Wallpaper App for certain devices such as dru. It also seemed
to stabilize the test case for some of the devices when opening the
productivity launcher again.

Tested on scarlet/dru and trogdor/lazor.

BUG=chromium:1353033, chromium:1360604
TEST=tast run DUT launcher.RecentApps*

Change-Id: I2c7425e5b489886edeab9f55670a04c50a472324