tast-tests: Close all open windows in drivefs fixture
When moving to the SWA version of Files app it doesn't reuse the
existing Files app window, so any subsequent launches will cause the
launching code to not identify the only window available.

This closes all the windows prior to running the Drivefs tests to ensure
the launching code launches fresh each time.

BUG=chromium:1360687
TEST=tast run filemanager.DrivefsGoogleDoc filemanager.DrivefsSearch

Change-Id: I2a414019a354b4d991891c9bc28c74a940297753