Adopt proto from Android 13
We are upgrading the .proto files from Android 11 to 13 in order to make
the arc tast tests based on `dumpsys --proto` work.
There are two main changes in the output for dumpsys activities that
affects the tests:
* The deprecation of WindowContainer in DisplayContent[ag/11579410]
* The introduction of TaskFragment[ag/14934423]
This commit updates dumpsys.go to adopt the changes.

BUG=b:245647771
TEST=`tast run $DUT arc.*` against R
TEST=`tast run $DUT arc.SplitView.*` against T and make sure getTaskInfo works

Change-Id: Ib9c8324d548d58585ebe6fdc3dda9f4c1af62c60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3886501
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Qijing Yao <yaoqq@google.com>
Reviewed-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Commit-Queue: Qijing Yao <yaoqq@google.com>