graphics: Change performance counter parse error to warning
We represent performance counters as time.Duration which is limited to
about 290 years and that should be plenty, but evidently the performance
counter "i915/vcs0-busy/" can report 18446743328000862208 ns which is
almost 585 years. That performance counter probably needs to be fixed,
but for now, passing test results will be valuable even without data for
that performance counter. The parse failure will be reported only as a
log message, no longer an actual error.

BUG=b:243564735
TEST=../tast/fast_build.sh -T

Change-Id: Ifc8c51eb08ec4df282146fb5c94e5fe1e10bea6a