tast-tests: Improve launcher.OpenExpandedView() reliability
For both the clamshell mode and tablet mode launcher, the launcher
widget is cached in an invisible state after it is hidden. When checking
to see if the launcher is open, explicitly check for a *visible* app
list item, as there may be invisible items in the view hierarchy.

BUG=chromium:1351225
TEST=tast run <dut> launcher.* has no new failures

Change-Id: I5404eb1a265f6ce3be6762443fbdafa350f02d0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3880607
Tested-by: James Cook <jamescook@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>