power: Add servov4 workaround to ShutdownWithCommandBatteryCutoff
The bug b/232284174 is currently tracking an issue where the servov4
will disconnect the charger. This is causing the test
power.ShutdownWithCommandBatteryCutoff to fail, since the AP is failing
to boot without AC power.

This CL works around this issue by rebooting the EC if it detects that
the AC charger is no longer connected. This appears to be fixing the
issue for now, allowing the test to pass on Skyrim.

BUG=b:217911928
BUG=b:232284174
TEST=tast run <IP> power.ShutdownWithCommandBatteryCutoff

Change-Id: I8cf72840ee0fc916e86326fbe20877365374a5fe