documentscanapi: Tighten button search criteria
The test searches for a button called "Scan" to trigger a
chrome.documentScan.scan call.  crrev.com/c/3871686 added a new launcher
button that has the same name, so this test fails flakily if that button
happens to be visible in the launcher.  Fix this by making sure that the
test only finds the "Scan" button inside our "Scanner Control" document.

BUG=chromium:1361234
TEST=tast run documentscanapi.Scan

Change-Id: Ie9cb6748fc0f7caeec647072775eb3a23bb52365