cellular: Fix SimConnect test for MBIM modems
Shill does not export services for the inactive eSIM slot if the iccid
of the eSIM slot is unknown. Chrome uses Hermes to switch slots in this
case. For inactive pSIMs, Shill exports a service  with
iccid="unknown-iccid", so that Chrome can connect to the pSIM.
See go/dual-sim-mbim for details. In order to predict the iccid of the
shill service, the test now starts on the eSIM and then switches to the
pSIM.

BUG=b:200813616
TEST=tast run $cr cellular.ShillCellularSimConnect

Change-Id: I2bc8139fb69c08cf4329db8f0532105b3685fb6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3852702
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>